refactor: 데이터 타입 수정
- 엔티티의 필드 타입에 맞게 Id 값의 타입은 Reference Type 사용
- 시간 또는 카운트 숫자의 데이터 경우 성능을 고려해서 Primitive Type 사용

+++ src/main/java/coffeemeet/server/matching/implement/MatchingQueueQuery.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class MatchingQueueQuery {
 
-  private static final long REDIS_IN_PIPELINE_OR_TRANSACTION = 0;
+  private static final int REDIS_IN_PIPELINE_OR_TRANSACTION = 0;
 
   private final RedisTemplate<String, Long> redisTemplate;
